도담이 설정 페이지 컴포넌트 분리
